backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.6.0
PyJWT==2.10.1
requests==2.31.0
openai==1.3.0
python-dateutil==2.8.2
bcrypt==4.1.1
gunicorn==21.2.0
mongomock==4.1.5
google-cloud-bigquery==3.13.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-auth==2.25.0
